+++ .eslintrc.yml
@@ -1,3 +1,7 @@
+
+parserOptions:
+ sourceType: 'module'
+
 env:
   es6: true
   browser: true
